compressor: rename Compressor/Decompressor to Encoder/Decoder, and rename CompressorManager to Compressor
